requirements.txt
PyPDF2==3.0.1
requests==2.31.0
gTTS==2.4.0
groq==0.3.2
python-dotenv==1.0.0
pydub==0.25.1